src/services/delivery/requirements.txt
grpcio
grpcio-tools
asyncpg
pydantic
passlib
bcrypt
email-validator